Reorganize a couple of the legacy helpers [npm node jenkinsplugin dynamic]
