+++ core/base-service/to-array.js
@@ -1,0 +1,11 @@
+'use strict'
+
+module.exports = function toArray(val) {
+  if (val === undefined) {
+    return []
+  } else if (Object(val) instanceof Array) {
+    return val
+  } else {
+    return [val]
+  }
+}
