+++ lib/badge-data.js
@@ -1,17 +1,8 @@
 'use strict'
 
 const coalesce = require('../core/base-service/coalesce')
+const toArray = require('../core/base-service/to-array')
 const { makeLogo } = require('./logos')
-
-function toArray(val) {
-  if (val === undefined) {
-    return []
-  } else if (Object(val) instanceof Array) {
-    return val
-  } else {
-    return [val]
-  }
-}
 
 function makeLabel(defaultLabel, overrides) {
   return `${
